Added RouteRankProcess unit test

+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/process/rank/RouteAndSurroundingFeatures.scala
@@ -202,7 +202,7 @@ class RouteAndSurroundingFeatures(val route: Route,
    */
   def evidenceOfMotion(tubeFeatures: Iterable[SimpleFeatureWithDateTimeAndKey],
                        boxFeatures: Iterable[SimpleFeatureWithDateTimeAndKey],
-                       routeDivisions: Double = RankingDefaults.defaultRouteDivisions): EvidenceOfMotion = {
+                       routeDivisions: Int = RankingDefaults.defaultRouteDivisions): EvidenceOfMotion = {
     val motionScores =
       boxFeatures
         .collect {
